Recovered from Eclipse.
git-svn-id: https://subversion.assembla.com/svn/cmosher@4903 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ GDiffView/src/nu/mine/mosher/GDiffView.java
@@ -170,7 +170,7 @@ public class GDiffView extends JFrame
                         beginSrc = copy.getRange().getBegin();
                         endSrc = copy.getRange().getEnd();
                         highlight(true,false);
-                        paneSrc.setCaretPosition(beginSrc);
+                        paneSrc.setCaretPosition((int)beginSrc);
                         beginTrg = copy.getTargetRange().getBegin();
                         endTrg = copy.getTargetRange().getEnd();
                         highlight(true,true);
